+++ src/main/java/net/geforcemods/securitycraft/ClientHandler.java
@@ -109,6 +109,7 @@ public class ClientHandler {
 				new ResourceLocation(SecurityCraft.MODID, "laser_block")
 		};
 		String[] mines = {
+				"blast_furnace",
 				"coal_ore",
 				"cobblestone",
 				"diamond_ore",
@@ -123,6 +124,7 @@ public class ClientHandler {
 				"nether_gold_ore",
 				"redstone_ore",
 				"sand",
+				"smoker",
 				"stone"
 		};
 		ResourceLocation[] furnaces = {