+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for v1.9.1 of SecurityCraft--------------------------
 
 - New: Password-protected Smoker and Password-protected Blast Furnace
+- New: Smoker and Blast Furnace mines
 - Change: The Password-protected Furnace now has a new look
 - Change: The Cage Trap is now reusable
 - Fix: Cannot access the recipe book for the Password-protected Furnace